refactor(prompt-area): extract clipboard-helpers and delete dead state
Moves the four clipboard/fragment utilities (serializeFragmentToPlainText,
serializeFragmentToSegments, parseSegmentsFromClipboard,
insertSegmentsAtCursor) from use-prompt-area-events.ts into a dedicated
clipboard-helpers.ts module. Private isRecord helper moves with them.

Undo/redo stays inline per plan scope-1 — a sub-hook would be a
pass-through wrapper with one caller. Adds a section comment above the
undoState ref documenting the ref-based identity invariant.

Also deletes the prevSegments ref (declared at line 103) and its 7 writes
— the value was never read anywhere in the repo. Behavior-identical;
removes 8 dead touch points.

use-prompt-area-events.ts: 661 lines → 405 lines.
Bumps registry entry + registry.test.ts to 19 files. All 593 tests pass.
